Added missing cast to avoid infinit recursion

+++ src/som/primitives/arithmetic/SubtractionPrim.java
@@ -45,6 +45,6 @@ public abstract class SubtractionPrim extends ArithmeticPrim {
 
   @Specialization
   public final double doDouble(final double left, final long right) {
-    return doDouble(left, right);
+    return doDouble(left, (double) right);
   }
 }
